fix(pkg/sshd): move utility func to common file
